Change evaluate_orbit to evaluate_orbits so that it can accept multiple orbits for a single set of observations
