+++ adam_core/orbit_determination/__init__.py
@@ -1,7 +1,5 @@
 # flake8: noqa: F401
-from .differential_correction import (
-    OrbitDeterminationObservations,
-    evaluate_orbit,
-    fit_least_squares,
-)
+from .differential_correction import fit_least_squares
+from .evaluate import OrbitDeterminationObservations, evaluate_orbits
 from .fitted_orbits import FittedOrbitMembers, FittedOrbits
+from .outliers import remove_lowest_probability_observation
